+++ com.unity.cinemachine/Runtime/Components/CinemachineSplineDolly.cs
@@ -117,8 +117,9 @@ namespace Cinemachine
         public DampingSettings Damping;
         
         /// <summary>Controls how automatic dollying occurs</summary>
-        [Tooltip("Controls how automatic dollying occurs.  A tracking target may be necessary to use this feature.")]
         [NoSaveDuringPlay]
+        [FoldoutWithEnabledButton]
+        [Tooltip("Controls how automatic dollying occurs.  A tracking target may be necessary to use this feature.")]
         public SplineAutoDolly AutomaticDolly;
 
         // State info for damping
@@ -134,8 +135,8 @@ namespace Cinemachine
             Damping.Position.y = Mathf.Clamp(Damping.Position.y, 0, 20);
             Damping.Position.z = Mathf.Clamp(Damping.Position.z, 0, 20);
             Damping.Angular = Mathf.Clamp(Damping.Angular, 0, 20);
-            if (AutomaticDolly.Implementation != null)
-                AutomaticDolly.Implementation.Validate();
+            if (AutomaticDolly.Method != null)
+                AutomaticDolly.Method.Validate();
         }
 
         void Reset()
@@ -146,7 +147,7 @@ namespace Cinemachine
             SplineOffset = Vector3.zero;
             CameraUp = CameraUpMode.Default;
             Damping = default;
-            AutomaticDolly.Implementation = null;
+            AutomaticDolly.Method = null;
         }
 
         /// <summary>Called when the behaviour is enabled.</summary>
@@ -154,6 +155,8 @@ namespace Cinemachine
         {
             base.OnEnable();
             RefreshRollCache();
+            if (AutomaticDolly.Method != null)
+                AutomaticDolly.Method.Reset();
         }
 
         /// <summary>True if component is enabled and has a spline</summary>
@@ -195,8 +198,8 @@ namespace Cinemachine
             }
 
             // Invoke AutoDolly algorithm to get new desired spline position
-            if (AutomaticDolly.Implementation != null)
-                splinePos = AutomaticDolly.Implementation.GetSplinePosition(
+            if (AutomaticDolly.Enabled && AutomaticDolly.Method != null)
+                splinePos = AutomaticDolly.Method.GetSplinePosition(
                     this, FollowTarget, Spline, splinePos, PositionUnits, deltaTime);
 
             // Apply damping in the spline direction