+++ com.unity.cinemachine/Runtime/Core/SplineAutoDolly.cs
@@ -10,9 +10,13 @@ namespace Cinemachine
     [Serializable]
     public struct SplineAutoDolly
     {
+        /// <summary>If set, will enable AutoDolly on a spline</summary>
+        [Tooltip("If set, will enable the selected automatic dolly along the spline")]
+        public bool Enabled;
+        
         /// <summary>This is the object that actually implements the AutoDolly</summary>
         [SerializeReference]
-        public ISplineAutoDolly Implementation;
+        public ISplineAutoDolly Method;
 
         /// <summary>
         /// Interface for procedural spline dolly.
@@ -22,6 +26,12 @@ namespace Cinemachine
         {
             /// <summary>Called from OnValidate() to validate the settings.</summary>
             void Validate();
+
+            /// <summary>Call this to reset any state information contained in the implementation.</summary>
+            void Reset();
+
+            /// <summary>Returns true if this implementation requires a tracking target.</summary>
+            bool RequiresTrackingTarget { get; }
 
             /// <summary>
             /// Compute the desired position on the spline.
@@ -50,6 +60,12 @@ namespace Cinemachine
 
             /// <summary>Called from OnValidate() to validate the settings.</summary>
             void ISplineAutoDolly.Validate() {}
+
+            /// <summary>This implementation does nothing.</summary>
+            void ISplineAutoDolly.Reset() {}
+
+            /// <summary>Returns true if this implementation requires a tracking target.</summary>
+            bool ISplineAutoDolly.RequiresTrackingTarget => false;
 
             /// <summary>
             /// Compute the desired position on the spline.
@@ -120,6 +136,12 @@ namespace Cinemachine
                 SearchResolution = Mathf.Max(SearchResolution, 1);
                 SearchIteration = Mathf.Max(SearchIteration, 1);
             }
+
+            /// <summary>This implementation does nothing.</summary>
+            void ISplineAutoDolly.Reset() {}
+
+            /// <summary>Returns true if this implementation requires a tracking target.</summary>
+            bool ISplineAutoDolly.RequiresTrackingTarget => true;
 
             /// <summary>
             /// Compute the desired position on the spline.